# Conflicts:
#	src/main/kotlin/ru/sber/kotlin/school/telegram/bot/model/User.kt
#	src/main/kotlin/ru/sber/kotlin/school/telegram/bot/model/Word.kt

+++ src/main/kotlin/ru/sber/kotlin/school/telegram/bot/repository/DictionaryRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository
 import ru.sber.kotlin.school.telegram.bot.model.Dictionary
 
 @Repository
-interface DictionaryRepository : JpaRepository<Dictionary, Int> {
+interface DictionaryRepository : JpaRepository<Dictionary, Long> {
 }
